move error type generation code inside the json protocol_generator, because the code it outputs is json specific
