+++ codegen/src/generator/mod.rs
@@ -6,8 +6,6 @@ use self::json::JsonGenerator;
 use self::query::QueryGenerator;
 use self::rest_json::RestJsonGenerator;
 
-use std::collections::HashMap;
-
 mod json;
 mod query;
 mod rest_json;
@@ -21,6 +19,10 @@ pub trait GenerateProtocol {
 
     fn generate_support_types(&self, _name: &str, _shape: &Shape, _service: &Service)
         -> Option<String> {
+        None
+    }
+
+    fn generate_error_types(&self, _service: &Service) -> Option<String> {
         None
     }
 }
@@ -45,7 +47,7 @@ fn generate<P>(service: &Service, protocol_generator: P) -> String where P: Gene
         client = generate_client(service, &protocol_generator),
         prelude = &protocol_generator.generate_prelude(service),
         types = generate_types(service, &protocol_generator),
-        error_types = generate_error_types(service).unwrap_or("".to_string()),
+        error_types = protocol_generator.generate_error_types(service).unwrap_or("".to_string()),
     )
 }
 
@@ -143,28 +145,7 @@ where P: GenerateProtocol {
     }).collect::<Vec<String>>().join("\n")
 }
 
-fn generate_error_types(service: &Service) -> Option<String> {
-    if service.typed_errors() {
 
-        // grab error type documentation for use with error enums in generated code
-        // botocore presents errors as structs.  we filter those out in generate_types.
-        let mut error_documentation = HashMap::new();
-
-        for (name, shape) in service.shapes.iter() {
-            if shape.exception() && shape.documentation.is_some() {
-                error_documentation.insert(name, shape.documentation.as_ref().unwrap());
-            }
-        }
-
-       Some(service.operations.iter()
-        .filter_map(|(_, operation)| generate_error_type(operation, &error_documentation) )
-        .collect::<Vec<String>>()
-        .join("\n")
-        )
-    } else {
-       None
-    }
-}
 
 fn generate_struct<P>(
     service: &Service,
@@ -235,106 +216,3 @@ impl Operation {
         format!("{}Error", self.name)
     }
 }
-
-pub fn generate_error_type(operation: &Operation, error_documentation: &HashMap<&String, &String>) -> Option<String> {
-
-    let error_type_name = operation.error_type_name();
-
-    Some(format!("
-        #[derive(Debug, PartialEq)]
-        pub enum {type_name} {{
-            {error_types}
-        }}
-
-        impl {type_name} {{
-            pub fn from_body(body: &str) -> {type_name} {{
-                match from_str::<Value>(body) {{
-                    Ok(json) => {{
-                        let error_type: &str = match json.find(\"__type\") {{
-                            Some(error_type) => error_type.as_string().unwrap_or(\"Unknown\"),
-                            None => \"Unknown\",
-                        }};
-
-                        match error_type {{
-                            {type_matchers}
-                        }}
-                    }},
-                    Err(_) => {type_name}::Unknown(String::from(body))
-                }}
-            }}
-        }}
-        impl From<AwsError> for {type_name} {{
-            fn from(err: AwsError) -> {type_name} {{
-                {type_name}::Unknown(err.message)
-            }}
-        }}
-        impl fmt::Display for {type_name} {{
-            fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {{
-                write!(f, \"{{}}\", self.description())
-            }}
-        }}
-        impl Error for {type_name} {{
-            fn description(&self) -> &str {{
-               match *self {{
-                   {description_matchers}
-               }}
-           }}
-       }}
-       ",
-       type_name = error_type_name,
-       error_types = generate_error_enum_types(operation, error_documentation).unwrap_or(String::from("")),
-       type_matchers = generate_error_type_matchers(operation).unwrap_or(String::from("")),
-       description_matchers = generate_error_description_matchers(operation).unwrap_or(String::from(""))))
-    }
-
-fn generate_error_enum_types(operation: &Operation, error_documentation: &HashMap<&String, &String>) -> Option<String> {
-    let mut enum_types: Vec<String> = Vec::new();
-
-    if operation.errors.is_some() {
-        for error in operation.errors.as_ref().unwrap().iter() {
-            enum_types.push(format!("\n///{}\n{}(String)",
-                error_documentation.get(&error.shape).unwrap_or(&&String::from("")),
-                error.idiomatic_error_name()));
-        }
-    }
-
-    enum_types.push("/// A validation error occurred.  Details from AWS are provided.\nValidation(String)".to_string());
-    enum_types.push("/// An unknown error occurred.  The raw HTTP response is provided.\nUnknown(String)".to_string());
-    Some(enum_types.join(","))
-}
-
-fn generate_error_type_matchers(operation: &Operation) -> Option<String> {
-    let mut type_matchers: Vec<String> = Vec::new();
-    let error_type = operation.error_type_name();
-
-    if operation.errors.is_some() {
-        for error in operation.errors.as_ref().unwrap().iter() {
-            type_matchers.push(format!("\"{error_shape}\" => {error_type}::{error_name}(String::from(body))",
-                error_shape = error.shape,
-                error_type = error_type,
-                error_name = error.idiomatic_error_name()))
-        }
-    }
-
-   type_matchers.push(format!("\"Validation\" => {error_type}::Validation(String::from(body))", error_type = error_type));
-   type_matchers.push(format!("_ => {error_type}::Unknown(String::from(body))",  error_type = error_type));
-   Some(type_matchers.join(","))
-}
-
-fn generate_error_description_matchers(operation: &Operation) -> Option<String> {
-    let mut type_matchers: Vec<String> = Vec::new();
-    let error_type = operation.error_type_name();
-
-    if operation.errors.is_some() {
-        for error in operation.errors.as_ref().unwrap().iter() {
-            type_matchers.push(format!("{error_type}::{error_shape}(ref cause) => cause",
-                error_type = operation.error_type_name(),
-                error_shape = error.idiomatic_error_name()))
-        }
-    }
-
-   type_matchers.push(format!("{error_type}::Validation(ref cause) => cause", error_type = error_type));
-   type_matchers.push(format!("{error_type}::Unknown(ref cause) => cause", error_type = error_type));
-   Some(type_matchers.join(","))
-}
-
